+++ rest/hosts_pool.go
@@ -191,9 +191,11 @@ func (s *Server) getHostInPool(w http.ResponseWriter, r *http.Request) {
 
 func (s *Server) listHostsInPool(w http.ResponseWriter, r *http.Request) {
 	filtersString := r.URL.Query()["filter"]
+
 	filters := make([]labelsutil.Filter, len(filtersString))
 	for i := range filtersString {
 		var err error
+
 		filters[i], err = labelsutil.CreateFilter(filtersString[i])
 		if err != nil {
 			writeError(w, r, newBadRequestError(err))